[FIX] l10n_mx_edi_landing: do not override customs number
When manually adding a customs number in draft invoices the value is
saved, but then overridden when the invoice is posted (confimed) with
the first value retrieved from the database.

The system should respect user choice as
some clients require this manual selection
regardless the FIFO or FEFO strategy.

opw-2616269

closes odoo/enterprise#21413

X-original-commit: ddba9117b781b17b77d455e6c74da8873c0eab3f
Signed-off-by: Josse Colpaert <jco@openerp.com>
Signed-off-by: agr-odoo <agr-odoo@users.noreply.github.com>