[FIX] account_bank_statement_import_camt: instructed amount in tx details
Before this commit, we were only looking for amount in tx details in
TransactionAmount or in CounterValueAmount elements. This can cause issue
if the amount in transaction details is only expressed in InstructedAmount
and if we have a batch containing several transactions, as we will retrieve
the total entry amount for each transaction instead. This commit adds the
management of instructed amount in transaction details to avoid this.

X-original-commit: 178daa742e7d9df5429109e89c21f1f8b76a7d29
Part-of: odoo/enterprise#21263