[FIX] account_intrastat,l10n_nl_intrastat: Add new transaction codes,
adapt intrastat declaration format

--- account_intrastat ---
New transaction codes have been specified by the European Commision in EU 2020/1197. The old codes cannot be removed as they have been used in previous transactions.

--- l10n_nl_intrastat ---

The format of the l10n_nl_intrastat declaration is changed according to the changes found here:
https://www.cbs.nl/en-gb/participants-survey/overzicht/businesses/onderzoek/international-trade-in-goods

The new codes are added in account_intrastat (see above) as this is an EU-wide change.

This means some changes to the intrastat declaration are required to accomodate the new two-digit transaction codes.

This means some changes to the intrastat declaration are required to accomodate the new two-digit transaction codes.

The user will be prevented from generating a monthly declaration that uses the old codes if the period is from January 2022 onward and prevented from generating a monthly declaration the new codes if the period is before January 2021.

The user will also be prevented from generating a declaration where both new and old codes are used in the same period.

The user is prevented not from using these codes in the wrong period but from generating a 'standard record layout (SRL)' declaration for submission using the IDEP+ web application. This is to allow some flexibility whilst maintaining the validity of the output file.

Task: 2635472
X-original-commit: 78d728a4a825d0eb6bb59442de1149f3b469d251
Part-of: odoo/enterprise#22781