[REF] web_cohort: pass model and callbacks props to renderer
By convention, views pass the model instance in props to the
renderer. Moreover, when the renderer needs to perform actions
outside of its natural scope, those callbacks have to be given in
props as well, s.t. the renderer can call them directly (instead
of triggering an event, listened by the parent).

X-original-commit: 525f314a0a13aec34a5a82c1baca0a406eaf0c95
Part-of: odoo/enterprise#21046