[FIX] l10n_be_hr_contract_salary: fix ui tour
The tour checking the contract salary in test_l10n_be_hr_payroll_account
fails since 14.4 due to the order of the values for the mobile advantage
not being correct. This commit sets the sequence explicitely on the
advantage values in order to make sure the order is the correct one.

TaskId-2651830

closes odoo/enterprise#21039

X-original-commit: 7013d20a8d0b5c16a5bc94947e5bb1537ae937d2
Signed-off-by: Kevin Baptiste <kba@odoo.com>
Signed-off-by: William Braeckman (wbr) <Williambraecky@users.noreply.github.com>