[REF] web_*: move noContentHelper to CohortView
Before this commit, the noContentHelper was handled at the renderer
level in cohort, whereas it's typically handled by the view itself.
This commit moves that logic up to the CohortView, thus closing
the gap between cohort and other views.

This commit also rework the way CohortModel's params are generated,
in a similar way that it is done in other views.

closes odoo/enterprise#21046

X-original-commit: c236e598eb41516d2fff05f29cfc61e4cde743d7
Related: odoo/odoo#76918
Signed-off-by: Géry Debongnie (ged) <ged@openerp.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>