[FIX] web_enterprise: fix barcode scanner
In commit, odoo/enterprise@9f5f38c67ff59b11d99000d7951f7a2703402f97
we have introduced the barcode scanner using native WebAPI Camera.

But inadvertently, the return value was wrongly adapted. This commit fixes it.

Steps to reproduce:
* Open the Odoo
* Go to the barcode app
* Click on "Tap to scan" button
* Try to scan a barcode => BUG the barcode is decoded but the functional
logic is not applied.

Note: this feature is only available on browser with a camera on the
device and in HTTPS only (e.g. Smartphone, laptop, ...).

closes odoo/enterprise#21142

X-original-commit: 9f20a21afdf47c528bf93983cd5efa976afba593
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>
Signed-off-by: rfr-odoo <rfr-odoo@users.noreply.github.com>