[FIX] sign: fixes validation banner unresponsive
In task-2475732, a new flow for signing was added, allowing sign items to be
added while signing. This new flow exposed a bug where the validation banner
would become disabled if you added a sign item and tried to sign without
filling it. This PR fixes this behavior, enabling the banner after an error
in the sign process. Also, when new sign items are added, a check is done to
decide if the banner should be shown or not.

task-2652338

closes odoo/enterprise#21567

X-original-commit: 216915d651eb0d64272ec1b3801e350cf4c4f2f7
Signed-off-by: Arnaud Joset <arj-odoo@users.noreply.github.com>