[FIX] {website_}appointment: fix auto published when duplicate
For appointment type, we don't want to copy the is_published field. When we
duplicate an appointment type, the category is copied. With the category copied
the default category is not used and so the appointment type was auto published
by the default_get method.

To avoid this, we set manually to False the is_published field when we copy
any category of appointment type. Also when we create/write an appointment
type (custom or work_hours), we auto publish it.

The goal is when we create a custom or work_hours to auto publish it on the
website (by the user flow or with the code). We keep the default behavior of
the copy of is_published for all categories.

task-2645163
ENT PR: odoo/enterprise#20901

X-original-commit: 9f9273ec17e5dc3fd23b0f97496e5374d165297c
Part-of: odoo/enterprise#22123