[FIX] appointment: create custom appointment from invitation menu
We should be able to create manually custom appointment from
`menu_calendar_appointment_type_custom_and_work_hours` menu with
unique type slots.

When creating a custom appointment type from this menu, we need a
way to know that the appointment type is custom to create unique slots.

The slot_type field of calendar.appointment.slot is now a computed and
stored field that depends from the category of the appointment type.
We store it because we target a stable version but the goal would be to
make it a non-stored computed field in master.

task-2645163
ENT PR: odoo/enterprise#20901

X-original-commit: c5cff900d8abdf6f469c6355cc6c745b0e8a3043
Part-of: odoo/enterprise#22123