[FIX] stock_barcode: inventory scan location
Before this commit, when the user scans a location in the barcode
inventory adjustment, that creates a new empty page for the scanned
location even if there is already an inventory page for this location.
It was due by the fact a page has a dest. location only for picking, so
for an inventory adjustment, the dest. location of a page is undefined,
and in case we change the page, we check if the wanted page exist (so we
display it) or not (so we create it).
In case the source or destination location isn't provided, we use the
default one, and as the default dest. location was `null` but the page's
one is `undefined` and those values, even if they are both falsy, aren't
strictly equals, it doesn't find the appropriate existing page.

task-2632818

X-original-commit: 872ad066302a8bc3a8961116253e91faa6033e17
Part-of: odoo/enterprise#21810