[FIX] account_bank_statement_import_camt: incorrect currency rate
In case the currency rate is provided in the counter value element, as CntrValAmt
is the countervalue in the account currency, we need to search the rate in the
target currency instead of the source currency. Otherwise, this will lead to an
incorrect conversion (division instead of multiplication). Some simplifications
were also made to remove unecessary rate_getters.

The existing multi currency test XML file was not correct and we need to fix the
rate to make it consistent with the actual related amounts in currencies.

closes odoo/enterprise#21263

X-original-commit: 6c1252ce29de982f5f6fabe9f9238763e9a7a21e
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>
Signed-off-by: Alex Tuyls <alt-odoo@users.noreply.github.com>