[FIX] documents_spreadsheet: throw human-readable error on failed name_get
When failing to fetch the display_name of a record in a pivot formula,
the error caught was simply stringified and cast as the error key of the
formula.

The error was not very explicit when read by a user ("Odoo Server
Error") but it was also causing issues during the tests : we were
catching the errors raised by the navigator while each navigator has its
own way to describe those errors.

For instance, in the tests, a TypeError was raised because we were trying
to access a property of an undefined object. One can then verify by
testing  `undefined.MyField` in several browser the following outputs:

Firefox 92.0 -> `TypeError: undefined has no properties`
Chrome < 93  -> `TypeError: Cannot read property 'display_name' of
undefined`
Chrome 93    -> `TypeError: Cannot read properties of undefined (reading
'displayName')`

Task 2653908

closes odoo/enterprise#21083

X-original-commit: 12d31921a5203afe524182a60c3ed838c48a21f6
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>