[IMP] social_youtube: display youtube thumbnails in (stream) post kanban views
This commits adds youtube thumbnails to social post kanban cards if the post
is linked to a youtube video. This thumbnails redirect on the related youtube
video.

For stream posts that contain a youtube link in their message body, a thumbnail
of the link youtube video is added to the kanban card.

To do both, new computed fields have been added:
- youtube_thumbnail_url
- youtube_video_url
Both field contains the url template for youtube video and thumbnail.
The advantage is that if the endpoint or part of the url changes, the update can
be done by code instead of having to update the templates.

Task-2516278
COM PR: odoo/odoo#76934
ENT PR: odoo/enterprise#21055