[FIX] web_dashboard: (legacy) fullscreen subviews
This commit makes sure that a legacy subview in a legacy dashboard can
still be opened in fullscreen. Why do we want this?

Let us assume we have a (new) dashboard view with a (new) graph subview
like in crm Pipeline. Let us add it to the dashboard application (board
module), reload, and go to board application. We have there a legacy
dashboard view with a legacy graph subview displayed (the dashboard app
still uses the old implementation of the reporting views). Expand the
graph subview causes a crash.

The crash occurs because the dashboard view exports a search model state
that contains only the key "DashboardModelExtension", so that the search
model of the (new) graph view displayed in fullscreen receives a state
that does not have the right form (it should contain at least the key
"ControlPanelModelExtension").

We fix this by stopping to export the dashboard controller state that
does not contain any pertinent information for its subviews. Indeed, the
legacy dashboard is now only used in the studio and dashboard apps and
 - in studio: a subview cannot be opened in fullscreen;
 - in dashboard app: a dashboard view does not install ControlPanelModelExtension
   or SearchPanelModelExtension (since withControlPanel and withSearchPanel
   are set to false), so that exportState in that case returns only
   information about the DashboardModelExtension that is installed only
   by the dashboard view.

closes odoo/enterprise#21156

X-original-commit: 9c117860fc026b5e47d63c41149cab47870cb5ed
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Co-authored-by: Lucas Perais <lpe@odoo.com>