[FIX] sale_subscription: check access_rule on write for payment token
In sale_subscription it is currently possible to write payment token records
via RPC call and no check is done to assert that the user has access to that
payment token. This means that it is possible to assign payment tokens from
other users to your subscriptions. This PR fixes this issue by checking the
access rules on the payment token. If an access error is raised, the write
call fails and the payment token is not changed.

task-2654115

closes odoo/enterprise#22017

X-original-commit: 66af6cd7af86488758cb0e2c05700c884e3442f9
Signed-off-by: Arnaud Joset <arj@odoo.com>
Signed-off-by: Leonardo Pavan Rocha <lpr@odoo.com>