[FIX] pos_blackbox_be: correctly compute discounted value
In the financial Z report, the discounts are not correctly computed,
because it was using an old behavior where the subtotal price was
computed, and it is not the case anymore.

So instead of changing the value of the line to compute the discount
apply, we are now getting the value without discount by computation.

OPW-2648233

closes odoo/enterprise#21145

X-original-commit: 4c925cb3eb620fdcda40ecc0698fa6609ce5ddad
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>