[FIX] appointment: fix python recurring slot test
There should be no slot when a slot is present in a futur monday
and that monday is also present in the first week of the next month.
Indeed we should check only the day for the month we display.
We also use the context_today to avoid possible issues with time
changes.

closes odoo/enterprise#21371

X-original-commit: 3e5f1efdcfaf0eedb732a33c3b53b392a6ffc56a
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>