[FIX] l10n_eu_oss_reports: disable tax closing for OSS reports
Automatic tax closing for OSS reports was mixed with the one made for VAT in the domestic country. The current implementation does not give us the possibility to clearly differentiate the closing moves made for OSS or VAT; so it is impossible to create an additional move for OSS.

Even more: OSS is made of two reports. Each of them should have its own closing entry.

This will be improved in master for v16. For v15, users will have to do their OSS closing by hand, as they used to before the introduction of automatic closing.

closes odoo/enterprise#21185

X-original-commit: 40455ba62c0bb053bb4d7484aba8372b9f79df34
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>