[IMP] industry_fsm: rename in '/my/tasks/<task_id>/*'
Before this commit, when the current user is in the '/my/tasks' route
and selects a task, the route becomes '/my/task/<task_id>'. There is no
reason to change the route and then adding the task id. For instance,
in the sales order route, we keep the route and we add the id of the SO
when the user has selected the SO that he wants to see more details.

This commit renames the route of task form view to keep the same route
and the same logic of SO and quotations routes for instance.

task-2648955

Community PR: odoo/odoo#82379

Closes #23315