[FIX] documents_account: sync payment term from documents actions button
Before this commit:

When we upload a file and set the contact with the newly created contact.
From Documents, you can click on the right menu on "create bills" or
"create customer invoice", but the generated record won't have its populated
payments term.

After this commit:

According to selected contact, payment terms will be populated.

Task-2645666

closes odoo/enterprise#28809

X-original-commit: 6f30a61ff4cb53a2d64b4363de71b67a4e27499e
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>