[FIX] sign: multiple opening of signature wizard
In sign, while signing a document, it is currently possible to click multiple
times on signature and initials sign items and this opens multiple signature
wizards. This PR fixes this issue by debouncing the click event handler to make
it sure it runs only once every 300ms at maximum.

task-2653071

closes odoo/enterprise#21842

X-original-commit: 648177f17e9af6ac20b0efb9506f5915c294403e
Signed-off-by: Arnaud Joset <arj-odoo@users.noreply.github.com>