[REF] *: adapt modules to action and view contexts
This commit follows the introduction of the view layout component as
well as the rewritten view templates. As such, it adapts all affected
templates to ensure that the XML overrides still work.

It also handles the adaptation of any component using the search
model's 'action' and/or 'view' key, now replaced by the environment
keys 'actionContext' and 'viewContext'.

Affected modules are:
- documents_spreadsheet
- web_enterprise
- website_sale_dashboard

X-original-commit: 3aabee9a033594da43d671d1b6fe5b12c54c7270
Part-of: odoo/enterprise#21295