[REF] web_*: namespace legacy views root classnames
*web_cohort,web_dashboard,web_studio

This commit is the counterpart of odoo/odoo#77294.

We namespace the classname of the root element of legacy views
that have been converted to owl, as both impementations are still
kept (even though, the legacy one is lazy-loaded).

That way, we prevent the scss rules of an implementation from
interfering with the other, and vice versa.

closes odoo/enterprise#21223

X-original-commit: eaedc801d68ab030cbe09997508c80e3f8e23ca7
Related: odoo/odoo#77323
Signed-off-by: Géry Debongnie (ged) <ged@openerp.com>