[FIX] currency_rate_live: code cleanup and robustness
Do not discard exceptions
=========================

Simply returning False makes it impossible to know what the error was
when it was raised since nothing is logged.
It is also hiding the real nature of the error one might not even have
thought of.
A bare except is also bad practice because it is catching more than
Exceptions

Add timeout and raise for status
================================

We don't want the cron pool to be saturated for longer than needed
because a distant server is not responding.
It is also cleaner to fail with a good type of Exception and message
thanks to `raise_for_status` instead of having keys missing in the json
dict later or whatnot.

Do not default provider only on company create
==============================================

Having a default + `create` extension won't change the value if you
change the country after having created the company, which might be the
case when setuping a new database for instance.

Require interval unit
=====================

Since the code doesn't make a difference between `manually` and `False`,
remove confusion by removing the ability to empty the field.

closes odoo/enterprise#21203

Signed-off-by: Josse Colpaert <jco@odoo.com>