[FIX] account_3way_match: Fix bills to pay filter
Since https://github.com/odoo/enterprise/pull/19531 The filter was wrong because the two conditions (sql request server-side and of the bills to pay filter) are not the same.

https://github.com/odoo/enterprise/blob/fdad6a7d1411a6f1f3685d6b40e2a81f30682b0c/account_3way_match/views/account_invoice_view.xml#L28
https://github.com/odoo/enterprise/blob/55998d8a05ae7c4d0eaee0b47874efa184a902f1/account_3way_match/models/account_journal_dashboard.py#L28

This means that we can have cases where:

It is displayed in the dashboard: 1 Bill To Pay
But once we click on it, the Bill To Pay filter doesn't show us any records

It can happen when the due date is equal to today.

OPW-2513075

closes odoo/enterprise#21329

X-original-commit: 32f83d89d54a805ffa5dd6ded289c965d3a611bf
Signed-off-by: William André (wan) <wan@odoo.com>