[IMP] industry_fsm_sale: allow portal user to see quotations
Before this commit, in project sharing feature, the portal user can see
the number of quotations via a stat button in the task view form but he
cannot click on this button because he could have no access to these
quotations.

This commit adds a new field called `portal_quotation_count` to display
the number of quotations the portal user can see.
If the portal user can see at least a quotation then he can click on
this button and he will be redirected to the quotes related to the task.

task-2648955

Community PR: odoo/odoo#82379

Closes #23315