[FIX] timesheet_grid: fix the invalid domain in timesheet header
Currently, when select the project from the timesheet header dropdown
then there is traceback due to invalid field in domain as allow_timesheet_timer
is removed from project.

So in this commit, replace the allow_timesheet_timer field by the
allow_timesheets so it will display all the project which has
timesheet enabled.

Related PR: #21206
Task-2657391

closes odoo/enterprise#21536

X-original-commit: 96d25c22ef95846aee67fe6db61177f7ce3d4cf4
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>
Signed-off-by: Xavier <xavierbol@users.noreply.github.com>