[IMP] {website_}appointment: adds new tab name & group-by filter
Currently, the tab name of the page listing the different calendar events
in the frontend (i.e: '/calendar') will contain the name of the active
layout. As example, if the administrator selects the card layout, the
tab name will be "Appointments Cards Layout" which is not ideal. To
make things clearer, we will set the page title of the list and card
design to "Appointment Types".

In this commit, we will also add a new preset filter that will allow the
user to group the records of the online appointments by types (in the
reporting section of the backend).

task-2646322

closes odoo/enterprise#21038

Related: odoo/odoo#76901
Signed-off-by: awa-odoo <awa-odoo@users.noreply.github.com>