[FIX] project_enterprise: create connectors on m2m grouping
Prior to this commit:
When grouping on m2m fields, only the first occurence of the record in the gantt
view was processed. The connectors were drawn whatever the number of fields in
the group by which could lead to crazy result, especially if groups were fold.

After this commit:
All the occurence of the task record are correctly processed. The connectors
will only be created if the group_by is made of one field or less.

task-2638949

closes odoo/enterprise#21497

X-original-commit: b9ca47ffa7fbefca0025bd102166e1dfbb5380bf
Signed-off-by: Kevin Baptiste <kba@odoo.com>
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>