[IMP] web_gantt: add progress bar in sidebar
This commit adds a progressbar representing the progress of a value
relatively to a max_value. Those values are given for a grouped field
which model has to define a gantt_progress_bar method.

This commit is a refactoring to take advantage of the implementation
given in the planning module and reused in project_enterprise.

PR: #21135
See odoo/odoo#77362

task-2646630