[FIX] timesheet_grid: fixed overtime calculation during time off
Currently in 15.0, while an employee is on leave, the timesheet app
considers it as overtime and due to that, it shows extra worked hours.
In actuality, the employee was on leave.

So in this commit, When employees put leaves than the Timesheet
consider it as leaves rather than overtime.

task-263559

closes odoo/enterprise#22233

X-original-commit: 6db3210c49c4f77cf89004af0e114f4562d7c322
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>