[FIX] hr_payroll: Prefill full time required hours
Purpose
=======

When creating a new calendar, for instance for a 9/10 contract, the
field full_time_required_hours is not prefilled.

If the HR sets it with a invalid value (for example 34.2 hours, which
is the supposed number of worked hours per week, this could lead to
inconsistencies, for example when computing the representation fees
proration, as the employee is considered to work full time.

To avoid mismatch, prefill the field with the real full time hours from
the company calendar, to prevent the HR setting it with the value they
think to be the correct one.

The field should remain editable, as the company calendar could be
fully open for companies working 24h per day for example.

Taskid: 2658157
X-original-commit: 674963e63458da2fe95783a8481d0875e4ca07bd
Part-of: odoo/enterprise#21286