[FIX] documents_spreadsheet: concurrent addition of pivot/list fails
When two users try to add a pivot/list concurrently, the second in list
will see his command rejected by the server. Unfortunately when
replaying his own command, he will override the datasource created by
the first user. This is due to the fact that the datasource addition is
handled by a ui plugin and is not adapted on replay.

Task 2658973

closes odoo/enterprise#21533

X-original-commit: 68fb891b90ef3179e5fabef789a347dd500a2e6d
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>