[FIX] sign: remove default role filter and rename default role to standard
A default role was added to sign in order to allow adding items while signing
documents without sign items. A filter currently exists to prevent the user
from choosing the default role when editing a template, however this filter
causes some tracebacks when we try to edit a template that contains the default
role. For this reason, it was decided that it was simpler to drop the filter and
allow users to select the default role, since it wouldn't affect the application.
Since the default role is now visible, it was decided to rename it to 'Standard'
not to mislead the users with the default word.

task-2657124

closes odoo/enterprise#21847

X-original-commit: 34b295fea09a02990607f1fa2c87511778361153
Signed-off-by: Arnaud Joset <arj@odoo.com>