[FIX] l10n_pe_edi: markupsafe does not end well in b64encode
Testing the PE edi in saas-14.5, we saw that the markupsafe object
was not well digested by the b64encode.

Furthermore an error in the code prevented the error to be shown
when there was an error in the response.

We improved some translation messages in the meantime while testing.

closes odoo/enterprise#21683

X-original-commit: f4367c54cb0a244291141438248dcefee0108119
Related: odoo/odoo#78352
Signed-off-by: William André (wan) <wan@odoo.com>