[MOV] industry_fsm_sale: move the invoice_count field
Before this commit, the `invoice_count` field of `project.task` model is
defined in `sale_project` module but it is not used in that module.

This commit moves the definition of this field in `industry_fsm_sale`
module since we use this field in this module.

task-2648955

Community PR: odoo/odoo#82379

Closes #23315