[FIX] pos_iot: Consider transaction cancelled on receiving error in TransactionResponse and not when the stage becomes blank
If a transaction request is sent to the terminal and followed by a
cancellation request, the cancellation request will get a cancellation
response. After that, the transaction response will be returned and it
will have the transaction result "Error". If you leave several seconds
between each of these operations, no errors will occur.

So the "normal" order of events when you start a transaction, cancel it,
and start a new one is:

T1: TransactionRequest
T1: CancelRequest
T1: CancelResponse
T1: TransactionResponse: Error
T2: TransactionRequest
T2: CancelRequest
T2: CancelResponse
T2: TransactionResponse: Error

If you quickly start a new transaction after cancelling, the order is
different:

T1: TransactionRequest
T1: CancelRequest
T1: CancelResponse
T2: TransactionRequest
T1: TransactionResponse: Error
T2: TransactionResponse: Error

In other words, starting the second transaction too quickly results in
it getting cancelled as well.

The TLV protocol defines a TerminalState field (section 4.2.2) that
describes whether a cancellation or a new transaction is possible. This
hasn't been implemented in the IngenicoDriver, but I did so as a test.
When the CancellationResponse is returned, the
"TransactionRequestPossible" flag (bit 1 in TerminalState) is still
false. That's why requesting a new transaction at that point fails.

It's only when the TransactionResponse is returned, indicating the
transaction has finished, that the TransactionRequestPossible flag
becomes true again.

I've experimented with several strategies to take the above into
account. The first was to try to prevent the problem at the level of
the IngenicoDriver, so if a cancellation was done and a new transaction
requested right after, it would wait until the old transaction had
finished and the new one was possible. This worked on the IngenicoDriver
level, but caused 2 problems on the POS side. The first is that
there's currently a 6 second timeout hard-coded on the POS side, which
would get triggered because the operation took too long as a result. The
time when the TransactionResponse is returned appears linked to the
keepalive interval (hard-coded to 5 seconds in the terminal). In theory
this could be worked around by increasing the timeout on the POS side.
Another issue is that the POS has no way of distinguishing the
TransactionResponses from the two different transactions. The
transaction id (or cid in the POS) remains the same. When you start the
second transaction in the POS, it'll wait for a response. If that
response has the error flag set, it'll assume the second transaction
failed. But if due to timing the response from the first transaction
arrives (which got cancelled and will also have the error flag set),
it'll assume the second transaction failed and will report that to the
user.

It's easy to propagate the TerminalState flags from the driver to the
POS, but it's not sufficient information for the POS to make a good
decision because it can't distinguish between the transactions. Changing
transaction id in the POS after cancelling a transaction seems like it
could be its own recipe for further synchronization issues.

So the cleanest way to deal with this seems to be to prevent starting
a new transaction from the POS UI side until it's certain the previous
one is finished. In the previous logic, the cancellation was considered
finished by the UI once the Stage field in the event data became blank.

The Stage field is the transaction stage from the TLV spec (section 4.4
.3). This field is for indicating which stage is displayed to the user
on the terminal. If we send a cancellation to the terminal, that stage
indeed becomes blank, but this doesn't mean the transaction is finished.
The only happens when a TransactionResponse is returned that has the
error field set.

So the solution I settled on here is to ignore the blank Stage field and
only consider the cancellation done when the stage is finished or the
error field is set. In theory we could also take into account the
TerminalState at this point if it were implemented in the
IngenicoDriver, but that doesn't seem to have added value at this point.

opw-2525010

closes odoo/enterprise#21710

X-original-commit: 9e83f681a0f655190d98421621027cb60692f5c4
Related: odoo/odoo#78414
Signed-off-by: Raf Geens <raf-odoo@users.noreply.github.com>
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>