[IMP] project_timesheet_forecast(_sale): generate timesheet entries from slot
Purpose of the commit is, the timesheets from the employees are be
pretty similar to that schedule. so an action to automatically convert
the shifts into timesheets would help the user gain a lot of time,
as the user would then only need to make small adjustments to his timesheets
instead of having to encode the whole thing from scratch.

So in this commit, add action to generate timesheet from slot based on
the number of allocated hours and the % of allocated time accross the
working days of resource calendar.
also on deletion of all the auto-generated entry allow to re-generate the
new entry from action.

task-2515308

closes odoo/enterprise#21376

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>