[FIX] sign: changes responsible of update log to token partner
In sign, when a document is edited while signing, an update log is created with
the partner who edited the template. If 'sign now' is used to sign a document in
the name of other person, and an update log is created, it was being created in
the name of the logged user who edited the template, however it should contain the
token owner.

task-2653064

closes odoo/enterprise#21358

Signed-off-by: Arnaud Joset <arj-odoo@users.noreply.github.com>