[IMP] test_mail_enterprise: add performance test for documents activities
We add documents in test_mail_enterprise dependencies in order to test features
and additional workload linked to document management in activities.

Add some query counters tests in test_mail_enterprise that installs a lot of
sub mail modules in both community and enterprise. Purpose is to have basics
for testing and some first query counter tests.

In this commit we test creation of activities linked to documents and upload.

Task-2657021

Part-of: odoo/enterprise#21200