[IMP] documents: split activity files and remove unnecessary fields definition
Purpose is to have activity type model defined in its own file to ease
code discovering.

We also remove unnecessary definition of ``default_user_id`` as it is defined
in ``mail`` module. Code is the same, no need to redefine twice the same code
and behavior.

Task-2657021

Part-of: odoo/enterprise#21200