[FIX] documents: wait for inspector in tests
The modified test selects records and checks the update in the
document inspector. It does it twice: once with a record linked
to a document, and one with a record not linked to anything. In
the first case, we wait for an extra tick s.t. the inspector is
updated, before doing the assertions. In the second case, we don't
wait, so the assertions are therefore always true (they are all
"negative" assertions, i.e. "containsNone").

closes odoo/enterprise#21460

Related: odoo/odoo#77861
Signed-off-by: Géry Debongnie (ged) <ged@openerp.com>