[FIX] web_studio: view_editor_manager correctly handles editing x2m
Before this commit, the View Editor Manager handled the edition of X2Many a bit roughly.
Because of the refactoring of the upper layer in wowl (87f2fe1b561757c3cb791ced17cee99ee6f8d6cb),
the VEM is always re-instanciated when clicking on the edition of a nested x2m view.

This caused the VEM to always instanciate the editor for the main view and only then started to update
itself for the edition of the x2m.

As a consequence, RPC were done too many times and the rendering was a bit slow.

After this commit, we instanciate the VEM in the right conditions from its start, allowing to have a smoother experience.

This comes at the price of a rather heavy fix, but it should prepare us for a future refactoring in OWL.

X-original-commit: e1182bfdfb39e69cf7735b49859d637fcb04b211
Part-of: odoo/enterprise#21527