[FIX] web_studio: navbar traceback on adapt
The navbar moves app sections under a "more" menu when they are overflowing.
Before this commit, Studio appended two buttons next to the app sections ones
but these two buttons are not true app sections.
This led to a traceback when the navbar ran its app sections overflow logic.

This commit places the two studio menu buttons next to
the app sections zone instead of inside of it.
This way, those studio buttons will never be part of the overflow logic
and will never be moved under the more menu.

X-original-commit: 8d9f95e055248e6a7c0fa2ca0e81f02b0a8dc9a2
Part-of: odoo/enterprise#21457