[FIX] web_enterprise: avoid scroll for background on Mobile
Before WOWL `.o_home_menu` was outside the `.o_action_manager`.
And since [1] a global CSS rule (in mail module) was added to
`.o_action_manager`:

```css
.o_action_manager {
    min-height: 0;
}
```

But after the commit [2] the `min-height` was removed, this as the
effect that `.o_action_manager` takes the whole height of it's content.
But the 'background' can't scroll.

This commit force the `.o_home_menu_scrollable` to has its  minimal size
at `0px`, so the scroll happens here and not in the parent.

Steps to reproduce:
* Open Odoo on Mobile
* Go to 'Home Menu' if not already in
* Scroll => BUG

Refs:
[1] odoo/odoo@3fea5b2136627c1d7e83b2a5e111952fe3eb10a1
[2] odoo/odoo@06081d2e8d86ec695f68c155b882d6a15a3fce9f

closes odoo/enterprise#21526

X-original-commit: 9a29f8e85cdcc1497c584ae191302b07a952df6f
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: rfr-odoo <rfr-odoo@users.noreply.github.com>