[IMP] account_invoice_extract_*: improve PO matching
The rules to match a PO were too strict, it didn't match
anything most of the time.

In this new version, we will match a PO even if it isn't a perfect
match and notify the user that he needs to manually check the
imported lines.

See task #2655372 for details on the new matching mechanisms.

closes odoo/enterprise#23090

X-original-commit: 9257fc63d6d0c79be5d8f2be464ef4dc35ce7901
Signed-off-by: Florian Daloze (fda) <fda@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>