[FIX] l10n_ae_hr_payroll: do not show EOS on payslip
End of service provision are not allowances, so its type has been
changed to company parts

The EOS Provision should not appear on payslips. The EOS Provision is
acting as a wallet that the company uses to pay the employees
terminated. Instead of the company expensing a huge amount in one
specific month where they paid end of service benefits to a former
employee, they would keep this provision entry (wallet) each month to
calculate the worth added to the employee's end of service benefit each
month they work, so at the end of employment I wouldn't expense the
whole amount in one month and overstate my expenses to that month.  This
rule shouldn't appear on the employee payslip because it is for
provision purposes (so that the company will be prepared for the worst
case scenario ). Employees shouldn't see it on their payslips

closes odoo/enterprise#21509

Related: odoo/upgrade#2891
X-original-commit: c873c2868a6623bdf1816b9d302109f005897631
Signed-off-by: Kevin Baptiste <kba@odoo.com>