[FIX] stock_barcode_picking_batch: suggested pack
Before this commit, if a package is scanned and assigned on a move line
of a batch, it will not be immediately suggested on the other same
picking's lines.
It was due to the suggested packages were only computed when the state
is created/refreshed. Now, it will also be done when a package is
assigned on a line.

task-2641528

closes odoo/enterprise#25471

Part-of: odoo/enterprise#20922
Signed-off-by: Arnold Moyaux <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>