[FIX] web_studio: click on the 'More' button is correctly rendered
Edit a view in studio in debug mode.
In the sidebar, click on the "View" tab, then click on the "More" button.

Before this commit, the ir.ui.view view on the record holding the arch was displayed in fullscreen.
This was because the do-action event propagated to the the window, where the legacyServiceProvider listens
to it, as a last resort API for doAction. The action was executed in the global webClient's context, and not in the studio
client action.

After this commit, the ir.ui.view is displayed in the studio client action, as wanted.

closes odoo/enterprise#21527

X-original-commit: e215d38425723d0ab95be77c2817570fbbff573f
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>